Correction 3

+++ 0x0D-preprocessor/3-function_like_macro.h
@@ -1,6 +1,6 @@
 #ifndef ABSS_H
 #define ABSS_H
 
-#define ABS(X) ((unsigned int)X >= 0UL ? ((unsigned int)X) : (unsigned int)-X)
+#define ABS(X) ((long)X >= 0L ? ((long)X) : (long)-X)
 
 #endif /*ABSS_H*/
